global_changes_v2 (rename algorithm to graph class and make_graph to backend)
